11th Commit -- worked on class files
Sub units now check if the main group contains the sub unit song before adding it to its list of songs. If it does contain the song, then the main group will remove the song before the sub unit adds it.
Also added an extra field for songs to store the japanese names. This will be based off of the ordering which the songs were inserted.